+++ CHANGELOG.md
@@ -49,6 +49,11 @@ This is a log of major user-visible changes in each phpMyFAQ release.
 - updated third party dependencies (Thorsten)
 - fixed bugs (Thorsten)
 
+### phpMyFAQ v4.1.8 - unreleased
+
+- made the database update resumable after a failed attempt on PostgreSQL (Thorsten)
+- fixed bugs (Thorsten)
+
 ### phpMyFAQ v4.1.7 - 2026-08-04
 
 - added CycloneDX SBOM (Software Bill of Materials) generation to the release packaging (Thorsten)